+++ lib/MIP/Cli/Mip/Analyse/Rd_rna.pm
@@ -91,7 +91,7 @@ sub run {
         salmon_quant_reference_genome       => [qw{ _salmon_quant_genome_dir }],
         star_aln_reference_genome           => [qw{ _star_genome_dir }],
         star_fusion_reference_genome        => [qw{ _star_fusion_genome_dir }],
-        transcript_annotation_file_endings  => [qw{ .refFlat .rrna.interval_list }],
+        transcript_annotation_file_endings  => [qw{ .bed .refflat .rrna.interval_list }],
     );
 
     mip_analyse(
@@ -145,6 +145,15 @@ sub _build_usage {
             documentation => q{Path to arriba protein domain file},
             is            => q{rw},
             isa           => Str,
+        )
+    );
+
+    option(
+        q{arriba_use_sample_id_as_display_name} => (
+            cmd_tags      => [q{Default: 0}],
+            documentation => q{Use sample id as display name for arriba fusion report},
+            is            => q{rw},
+            isa           => Bool,
         )
     );
 
@@ -495,15 +504,6 @@ q{Default: grch37_dbsnp_-138-.vcf, grch37_1000g_indels_-phase1-.vcf, grch37_mill
     );
 
     option(
-        q{rseqc_transcripts_file} => (
-            cmd_tags      => [q{Rseqc transcripts file: Format: GTF}],
-            documentation => q{Input for rseqc to build transcript bed format file},
-            is            => q{rw},
-            isa           => Str,
-        )
-    );
-
-    option(
         q{gatk_haplotypecaller} => (
             cmd_tags      => [q{Analysis recipe switch}],
             documentation => q{Variant discovery using GATK HaplotypeCaller},
@@ -571,6 +571,42 @@ q{Default: BaseQualityRankSumTest, ChromosomeCounts, Coverage, DepthPerAlleleByS
             documentation => q{Picardtools markduplicates optical duplicate distance},
             is            => q{rw},
             isa           => Int,
+        )
+    );
+
+    option(
+        q{megafusion_ar} => (
+            cmd_tags      => [q{Analysis recipe switch}],
+            documentation => q{Merge fusion calls from different callers},
+            is            => q{rw},
+            isa           => enum( [ 0, 1, 2 ] ),
+        )
+    );
+
+    option(
+        q{megafusion_callers} => (
+            cmd_tags      => [q{Default: arriba_ar star_fusion}],
+            documentation => q{Fusion caller recipes to fuse into a vcf},
+            is            => q{rw},
+            isa           => ArrayRef,
+        )
+    );
+
+    option(
+        q{megafusion_arriba_config} => (
+            cmd_tags      => [q{Default: Path inside MegaFusion container}],
+            documentation => q{MegaFusion config for Arriba calls},
+            is            => q{rw},
+            isa           => Str,
+        )
+    );
+
+    option(
+        q{megafusion_star_fusion_config} => (
+            cmd_tags      => [q{Default: Path inside MegaFusion container}],
+            documentation => q{MegaFusion config for Star Fusion calls},
+            is            => q{rw},
+            isa           => Str,
         )
     );
 
@@ -737,6 +773,16 @@ q{Regular expression file containing the regular expression to be used for each 
             isa           => Bool,
         )
     );
+
+    option(
+        q{qccollect_store_metrics_outfile} => (
+            cmd_tags      => [q{Default: {outdata_dir}/{case_id}_metrics_deliverables.yaml}],
+            documentation => q{File containing metrics from this analysis run},
+            is            => q{rw},
+            isa           => Str,
+        )
+    );
+
     option(
         q{multiqc_ar} => (
             cmd_tags      => [q{Analysis recipe switch}],
@@ -822,6 +868,15 @@ q{Regular expression file containing the regular expression to be used for each 
             documentation => q{Strandedness of library},
             is            => q{rw},
             isa => ArrayRef [ enum( [qw{ unstranded forward_stranded reverse_stranded }] ), ],
+        )
+    );
+
+    option(
+        q{svdb_merge_fusion} => (
+            cmd_tags      => [q{Analysis recipe switch}],
+            documentation => q{Merge sample fusion vcfs to a case vcf},
+            is            => q{rw},
+            isa           => enum( [ 0, 1, 2 ] ),
         )
     );