+++ lib/MIP/Program/Megafusion.pm
@@ -1,0 +1,124 @@
+package MIP::Program::Megafusion;
+
+use 5.026;
+use Carp;
+use charnames qw{ :full :short };
+use English qw{ -no_match_vars };
+use open qw{ :encoding(UTF-8) :std };
+use Params::Check qw{ allow check last_error };
+use utf8;
+use warnings;
+use warnings qw{ FATAL utf8 };
+
+## CPANM
+use autodie qw{ :all };
+use Readonly;
+
+## MIPs lib/
+use MIP::Constants qw{ $SPACE };
+use MIP::Environment::Executable qw{ get_executable_base_command };
+use MIP::Unix::Standard_streams qw{ unix_standard_streams };
+use MIP::Unix::Write_to_file qw{ unix_write_to_file };
+
+BEGIN {
+    require Exporter;
+    use base qw{ Exporter };
+
+    # Set the version for version checking
+    # Functions and variables which can be optionally exported
+    our @EXPORT_OK = qw{ megafusion };
+}
+
+Readonly my $BASE_COMMAND => q{MegaFusion.py};
+
+sub megafusion {
+
+## Function : Perl wrapper for MegaFusion
+## Returns  : @commands
+## Arguments: $config_file_path       => Json config file path
+##          : $filehandle             => Filehandle to write to
+##          : $infile_path            => Infile path
+##          : $sample_id              => Sample name
+##          : $stderrfile_path        => Stderrfile path
+##          : $stderrfile_path_append => Append stderr info to file path
+##          : $stdoutfile_path        => Stdoutfile path
+
+    my ($arg_href) = @_;
+
+    ## Flatten argument(s)
+    my $config_file_path;
+    my $filehandle;
+    my $infile_path;
+    my $sample_id;
+    my $stderrfile_path;
+    my $stderrfile_path_append;
+    my $stdoutfile_path;
+
+    my $tmpl = {
+        config_file_path => {
+            defined     => 1,
+            required    => 1,
+            store       => \$config_file_path,
+            strict_type => 1,
+        },
+        filehandle => {
+            store => \$filehandle,
+        },
+        infile_path => {
+            defined     => 1,
+            required    => 1,
+            store       => \$infile_path,
+            strict_type => 1,
+        },
+        sample_id => {
+            store       => \$sample_id,
+            strict_type => 1,
+        },
+        stderrfile_path => {
+            store       => \$stderrfile_path,
+            strict_type => 1,
+        },
+        stderrfile_path_append => {
+            store       => \$stderrfile_path_append,
+            strict_type => 1,
+        },
+        stdoutfile_path => {
+            store       => \$stdoutfile_path,
+            strict_type => 1,
+        },
+    };
+
+    check( $tmpl, $arg_href, 1 ) or croak q{Could not parse arguments!};
+
+    my @commands = ( get_executable_base_command( { base_command => $BASE_COMMAND, } ), );
+
+    push @commands, q{--json} . $SPACE . $config_file_path;
+
+    push @commands, q{--fusion} . $SPACE . $infile_path;
+
+    if ($sample_id) {
+
+        push @commands, q{--sample} . $SPACE . $sample_id;
+    }
+
+    push @commands,
+      unix_standard_streams(
+        {
+            stderrfile_path        => $stderrfile_path,
+            stderrfile_path_append => $stderrfile_path_append,
+            stdoutfile_path        => $stdoutfile_path,
+        }
+      );
+
+    unix_write_to_file(
+        {
+            commands_ref => \@commands,
+            filehandle   => $filehandle,
+            separator    => $SPACE,
+
+        }
+    );
+    return @commands;
+}
+
+1;