+++ t/analysis_genebody_coverage.t
@@ -30,8 +30,7 @@ BEGIN {
 ### Check all internal dependency modules and imports
 ## Modules with import
     my %perl_module = (
-        q{MIP::Recipes::Analysis::Genebody_coverage} =>
-          [qw{ analysis_genebody_coverage }],
+        q{MIP::Recipes::Analysis::Genebody_coverage} => [qw{ analysis_genebody_coverage }],
         q{MIP::Test::Fixtures} => [qw{ test_add_io_for_recipe test_log test_mip_hashes }],
     );
 
@@ -65,7 +64,7 @@ $active_parameter{recipe_core_number}{$recipe_name} = 1;
 $active_parameter{recipe_time}{$recipe_name}        = 1;
 my $sample_id = $active_parameter{sample_ids}[0];
 $active_parameter{star_aln_reference_genome} = q{genome.fasta};
-$active_parameter{rseqc_transcripts_file}    = q{transcripts.bed};
+$active_parameter{transcript_annotation}     = q{transcripts.gtf};
 
 my %file_info = test_mip_hashes(
     {
@@ -84,11 +83,11 @@ my %parameter = test_mip_hashes(
 
 test_add_io_for_recipe(
     {
-        file_info_href    => \%file_info,
-        id                => $sample_id,
-        parameter_href    => \%parameter,
-        recipe_name       => $recipe_name,
-        step              => q{bam},
+        file_info_href => \%file_info,
+        id             => $sample_id,
+        parameter_href => \%parameter,
+        recipe_name    => $recipe_name,
+        step           => q{bam},
     }
 );